requirements.txt
openai
dotenv
tiktoken
google-genai
httpx